refactor: folderwindow.tsx
